chore: Modularise feedback endpoint
